chore: resolve dayjs conflict
