+++ components/date-picker/__tests__/DatePicker.test.tsx
@@ -1,11 +1,15 @@
+import dayjs from 'dayjs';
+import 'dayjs/locale/mk'; // to test local in 'prop locale should works' test case
+import customParseFormat from 'dayjs/plugin/customParseFormat';
 import MockDate from 'mockdate';
-import moment from 'moment';
 import React from 'react';
 import type { TriggerProps } from 'rc-trigger';
 import { fireEvent, render } from '../../../tests/utils';
 import DatePicker from '..';
 import focusTest from '../../../tests/shared/focusTest';
 import type { PickerLocale } from '../generatePicker';
+
+dayjs.extend(customParseFormat);
 
 let triggerProps: TriggerProps;
 
@@ -29,7 +33,7 @@ describe('DatePicker', () => {
   focusTest(DatePicker, { refFocus: true });
 
   beforeEach(() => {
-    MockDate.set(moment('2016-11-22').valueOf());
+    MockDate.set(dayjs('2016-11-22').valueOf());
   });
 
   afterEach(() => {
@@ -77,13 +81,13 @@ describe('DatePicker', () => {
         placeholder: 'Избор на час',
       },
     };
-    const birthday = moment('2000-01-01', 'YYYY-MM-DD');
+    const birthday = dayjs('2000-01-01', 'YYYY-MM-DD');
     const wrapper = render(<DatePicker open locale={locale as PickerLocale} value={birthday} />);
     expect(Array.from(wrapper.container.children)).toMatchSnapshot();
   });
 
   it('disabled date', () => {
-    const disabledDate = (current: any) => current && current < moment().endOf('day');
+    const disabledDate = (current: any) => current && current < dayjs().endOf('day');
     const wrapper = render(<DatePicker disabledDate={disabledDate} open />);
     expect(Array.from(wrapper.container.children)).toMatchSnapshot();
   });
@@ -96,7 +100,7 @@ describe('DatePicker', () => {
   it('showTime={{ showHour: true, showMinute: true }}', () => {
     const { container } = render(
       <DatePicker
-        defaultValue={moment()}
+        defaultValue={dayjs()}
         showTime={{ showHour: true, showMinute: true }}
         format="YYYY-MM-DD"
         open
@@ -118,7 +122,7 @@ describe('DatePicker', () => {
   it('showTime={{ showHour: true, showSecond: true }}', () => {
     const { container } = render(
       <DatePicker
-        defaultValue={moment()}
+        defaultValue={dayjs()}
         showTime={{ showHour: true, showSecond: true }}
         format="YYYY-MM-DD"
         open
@@ -140,7 +144,7 @@ describe('DatePicker', () => {
   it('showTime={{ showMinute: true, showSecond: true }}', () => {
     const { container } = render(
       <DatePicker
-        defaultValue={moment()}
+        defaultValue={dayjs()}
         showTime={{ showMinute: true, showSecond: true }}
         format="YYYY-MM-DD"
         open
@@ -160,7 +164,7 @@ describe('DatePicker', () => {
   });
   it('showTime should work correctly when format is custom function', () => {
     const { container } = render(
-      <DatePicker defaultValue={moment()} showTime format={val => val.format('YYYY-MM-DD')} open />,
+      <DatePicker defaultValue={dayjs()} showTime format={val => val.format('YYYY-MM-DD')} open />,
     );
     const fuousEvent = () => {
       fireEvent.focus(container.querySelector('input')!);
@@ -174,7 +178,7 @@ describe('DatePicker', () => {
 
   it('12 hours', () => {
     const { container } = render(
-      <DatePicker defaultValue={moment()} showTime format="YYYY-MM-DD HH:mm:ss A" open />,
+      <DatePicker defaultValue={dayjs()} showTime format="YYYY-MM-DD HH:mm:ss A" open />,
     );
     expect(container.querySelectorAll('.ant-picker-time-panel-column').length).toBe(4);
     expect(
@@ -201,7 +205,7 @@ describe('DatePicker', () => {
 
   it('24 hours', () => {
     const { container } = render(
-      <DatePicker defaultValue={moment()} showTime format="YYYY-MM-DD HH:mm:ss" open />,
+      <DatePicker defaultValue={dayjs()} showTime format="YYYY-MM-DD HH:mm:ss" open />,
     );
     expect(container.querySelectorAll('.ant-picker-time-panel-column').length).toBe(3);
     expect(
@@ -236,8 +240,8 @@ describe('DatePicker', () => {
   });
 
   it('DatePicker.RangePicker with defaultPickerValue and showTime', () => {
-    const startDate = moment('1982-02-12');
-    const endDate = moment('1982-02-22');
+    const startDate = dayjs('1982-02-12');
+    const endDate = dayjs('1982-02-22');
 
     const { container } = render(
       <DatePicker.RangePicker defaultPickerValue={[startDate, endDate]} showTime open />,
