+++ components/date-picker/__tests__/RangePicker.test.tsx
@@ -1,5 +1,5 @@
-import moment from 'moment';
-import type { RangeValue } from 'rc-picker/lib/interface';
+import dayjs from 'dayjs';
+import customParseFormat from 'dayjs/plugin/customParseFormat';
 import React from 'react';
 import DatePicker from '..';
 import focusTest from '../../../tests/shared/focusTest';
@@ -7,6 +7,8 @@ import { render, resetMockDate, setMockDate } from '../../../tests/utils';
 import enUS from '../locale/en_US';
 
 import { closePicker, openPicker, selectCell } from './utils';
+
+dayjs.extend(customParseFormat);
 
 const { RangePicker } = DatePicker;
 
@@ -23,7 +25,7 @@ describe('RangePicker', () => {
 
   // issue: https://github.com/ant-design/ant-design/issues/5872
   it('should not throw error when value is reset to `[]`', () => {
-    const birthday = moment('2000-01-01', 'YYYY-MM-DD');
+    const birthday = dayjs('2000-01-01', 'YYYY-MM-DD');
     const wrapper1 = render(<RangePicker value={[birthday, birthday]} open />);
     const wrapper2 = render(<RangePicker value={[] as unknown as null} open />);
 
@@ -54,7 +56,7 @@ describe('RangePicker', () => {
   // https://github.com/ant-design/ant-design/issues/13302
   describe('in "month" mode, when the left and right panels select the same month', () => {
     it('the cell status is correct', () => {
-      let rangePickerValue = [] as unknown as RangeValue<any>;
+      let rangePickerValue: dayjs.Dayjs[] = [] as any;
       class Test extends React.Component {
         state = { value: null };
 
@@ -65,7 +67,7 @@ describe('RangePicker', () => {
               mode={['month', 'month']}
               onPanelChange={value => {
                 this.setState({ value });
-                rangePickerValue = value;
+                rangePickerValue = value as any;
               }}
             />
           );
@@ -79,7 +81,7 @@ describe('RangePicker', () => {
       selectCell(wrapper, 'Feb');
       closePicker(wrapper, 1);
 
-      const [start, end] = rangePickerValue as [moment.Moment, moment.Moment];
+      const [start, end] = rangePickerValue;
 
       expect(start.isSame(end, 'date')).toBeTruthy();
     });
@@ -91,8 +93,8 @@ describe('RangePicker', () => {
         <RangePicker
           open
           ranges={{
-            Today: [moment(), moment()],
-            'This Month': [moment().startOf('month'), moment().endOf('month')],
+            Today: [dayjs(), dayjs()],
+            'This Month': [dayjs().startOf('month'), dayjs().endOf('month')],
           }}
         />,
       );
